+++ airflow-core/tests/unit/cli/commands/test_dag_command.py
@@ -1082,6 +1082,7 @@ class TestCliDags:
 
         mock_dagbag.assert_called_once_with(
             bundle_path=TEST_DAGS_FOLDER,
+            bundle_import_root=TEST_DAGS_FOLDER,
             dag_folder=TEST_DAGS_FOLDER,
             bundle_name="testing",
         )
@@ -1103,6 +1104,7 @@ class TestCliDags:
 
         mock_dagbag.assert_called_once_with(
             bundle_path=TEST_DAGS_FOLDER,
+            bundle_import_root=TEST_DAGS_FOLDER,
             dag_folder=str(dag_file),
             bundle_name="testing",
         )
@@ -1134,6 +1136,7 @@ class TestCliDags:
 
         mock_dagbag.assert_called_once_with(
             bundle_path=TEST_DAGS_FOLDER,
+            bundle_import_root=TEST_DAGS_FOLDER,
             dag_folder=str(dag_file),
             bundle_name="testing",
         )