+++ task-sdk/tests/task_sdk/execution_time/test_task_runner.py
@@ -308,6 +308,7 @@ def test_parse_dag_bag(mock_dagbag, test_dags_dir: Path, make_ti_context):
         safe_mode=False,
         load_op_links=False,
         bundle_path=test_dags_dir,
+        bundle_import_root=test_dags_dir,
         bundle_name="my-bundle",
     )